requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
exceptiongroup==1.1.3
fastapi==0.103.0
idna==3.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
typing_extensions==4.7.1
